% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModels.R
\name{linearModels}
\alias{linearModels}
\title{Fitting Linear Models}
\usage{
linearModels(data, x, y)
}
\arguments{
\item{data}{The data set for multiple linear regression. It should be in the form of a 2d matrix}

\item{x}{A string array of all the predictors' names.}

\item{y}{A single string of the response.}
}
\value{

}
\description{
linearModels is a mimic of lm() that is used to fit linear models.
}
\examples{
mt = mtcars
linearModels(mtcars, "mpg", "cyl")
}
