% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\docType{data}
\name{LinearRegression}
\alias{LinearRegression}
\title{LinearRegression}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LinearRegression
}
\description{
LinearRegression
}
\examples{

library(linearmodels)

x <- matrix(runif(100), ncol = 2)
y <- matrix(2*x[,1] + 3*x[, 2] + rnorm(50), ncol = 1)

linear_regression <- LinearRegression$new()
linear_regression$fit(x, y)
linear_regression$predict(x)
}
\keyword{datasets}
