% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticLossEarlyStopping.R
\name{LMLogisticLossEarlyStoppingCV}
\alias{LMLogisticLossEarlyStoppingCV}
\title{LMLogisticLossEarlyStoppingCV}
\usage{
LMLogisticLossEarlyStoppingCV(X.mat, y.vec, fold.vec = NULL,
  max.iterations = 100, step.size = 0.35)
}
\arguments{
\item{X.mat}{An unscaled feature matrix [n x p]}

\item{y.vec}{A vector of predictions [p x 1]}

\item{fold.vec}{A vector of FoldID to pass for the cross-validation data split [p x 1]}

\item{max.iterations}{The number of times to step through your gradient descent [scalar]}

\item{step.size}{The size of the step to take during gradient descent [scalar]}
}
\value{
A list comprised of: mean.valid.vec [max.iterations x 1], selected.steps, w.vec [n x 1], and a function predict()
}
\description{
LMLogisticLossEarlyStoppingCV
}
\examples{
data(spam, package="ElemStatLearn")
X.mat.binary <- spam[,1:ncol(spam) -1]
y.vec.label <- spam[,'spam']
y.vec.binary <- ifelse(y.vec.label == 'spam', 1, 0)
max.iterations=50
step.size=0.001
fitLog <- LMLogisticLossEarlyStoppingCV(X.mat.binary, y.vec.binary, NULL, max.iterations, step.size)
fitLog$predict(x.mat.binary)
fitLog$mean.valid.vec
fitLog$selected.steps
fitLog$w.vec
}
