% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_Models.r
\name{LMLogisticLossL2}
\alias{LMLogisticLossL2}
\title{L2 Logistic loss}
\usage{
LMLogisticLossL2(X.scaled.mat, y.vec, penalty, opt.thresh,
  initial.weight.vec, step.size)
}
\arguments{
\item{X.scaled.mat}{numeric scaled input feature matrix [n x p]}

\item{y.vec}{numeric input label vector [n]}

\item{penalty}{numeric value for the penalty on weights}

\item{opt.thresh}{numeric value compare for convergence}

\item{initial.weight.vec}{numeric 0 vector [p]}

\item{step.size}{numeric learning rate}
}
\value{
numeric scaled optimal weight vector for the given penalty [p]
}
\description{
This function attempts to minimize the Loss function log[1 + exp(-YXw)]
}
\examples{
data(spam, package="ElemStatLearn")
X.mat.binary <- spam[,1:ncol(spam) -1]
y.vec.label <- spam[,'spam']
y.vec.binary <- ifelse(y.vec.label == 'spam', 1, 0 )
penalty <- 5
mean <- apply(X.mat.binary, 2, mean)
sd <- apply(X.mat.binary, 2, sd)
X.scaled.mat <- scale(X.mat.binary, center = mean, scale = sd)
opt.thresh <- 0.005
step.size = 0.1
initial.weight.vec <- integer(ncol(X.scaled.mat))
LMLogisticLossL2( X.scaled.mat, y.vec.binary, penalty, opt.thresh, initial.weight.vec, step.size )
}
