% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_Models.r
\name{LMLogisticLossL2CV}
\alias{LMLogisticLossL2CV}
\title{L2 Logistic CV}
\usage{
LMLogisticLossL2CV(X.mat, y.vec, fold.vec = NULL, penalty.vec)
}
\arguments{
\item{X.mat}{numeric input feature matrix [n x p]}

\item{y.vec}{numeric input label vector [n]}

\item{fold.vec}{numeric fold ids for each observation in X.mat [n]}

\item{penalty.vec}{numeric vector of decending penalties for weights [m]}
}
\value{
list containing the mean validation loss, mean training loss vector, pentalty vector, optimal penalty, optimal weight vector using optimal penalty, predict function
}
\description{
Cross validation using L2 Logistic regularization
}
\examples{
data(spam, package="ElemStatLearn")
X.mat.binary <- spam[,1:ncol(spam) -1]
y.vec.label <- spam[,'spam']
y.vec.binary <- ifelse(y.vec.label == 'spam', 1, 0 )
penalty.vec <- seq(10, 1, by=-0.5)
LMLogisticLossL2CV(X.mat.binary, y.vec.binary, fold.vec, penalty.vec)
}
