% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_Models.r
\name{LMLogisticLossL2penalties}
\alias{LMLogisticLossL2penalties}
\title{L2 Square Loss}
\usage{
LMLogisticLossL2penalties(X.mat, y.vec, penalty.vec)
}
\arguments{
\item{X.mat}{numeric input feature matrix [n x p]}

\item{y.vec}{numeric input label vector [n]}

\item{penalty.vec}{numeric vector of decending penalties for weights [m]}
}
\value{
numeric weight matrix with the optimal weight for every penalty value and the intercept term augmented to the top row of each vector [ p + 1 x m]
}
\description{
returns a weight matrix where every vector is the optimal weight for a given penalty
}
\examples{
data(spam, package="ElemStatLearn")
X.mat.binary <- spam[,1:ncol(spam) -1]
y.vec.label <- spam[,'spam']
y.vec.binary <- ifelse(y.vec.label == 'spam', 1, 0 )
penalty.vec <- seq(10, 1, by=-0.5)
LMLogisticLossL2penalties(X.mat.binary, y.vec.binary, penalty.vec)
}
