% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquareLossEarlyStopping.r
\name{LMSquareLossEarlyStoppingCV}
\alias{LMSquareLossEarlyStoppingCV}
\title{LMSquareLossEarlyStoppingCV}
\usage{
LMSquareLossEarlyStoppingCV(X.mat, y.vec, fold.vec = NULL,
  max.iterations = 100, step.size = 0.35)
}
\arguments{
\item{X.mat}{An unscaled feature matrix [n x p]}

\item{y.vec}{A vector of predictions [p x 1]}

\item{fold.vec}{A vector of FoldID to pass for the cross-validation data split [p x 1]}

\item{max.iterations}{The number of times to step through your gradient descent}

\item{step.size}{The size of the step to take during gradient descent}
}
\value{
A list comprised of: mean.valid.vec [max.iterations x 1], selected.steps, w.vec [max.iterations x p], and a function predict()
}
\description{
LMSquareLossEarlyStoppingCV
}
\examples{
X.mat.binary <- data[,-1]
y.vec.binary <-data[,1]
max.iterations=10
step.size=0.35
fitLog <- LMLogisticLossEarlyStoppingCV(X.mat.binary, y.vec.binary, NULL, max.iterations, step.size)
fitLog$predict(x.unsc.mat)
fitLog$mean.valid.vec
fitLog$selected.steps
fitLog$w.vec
}
