% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquareLossEarlyStopping.r
\name{LMSquareLossIterations}
\alias{LMSquareLossIterations}
\title{LMSquareLossIterations}
\usage{
LMSquareLossIterations(x.unsc.mat, y.vec, max.iterations, step.size)
}
\arguments{
\item{x.unsc.mat}{An unscaled feature matrix [n x p]}

\item{y.vec}{A vector of predictions [p x 1]}

\item{max.iterations}{The number of times to step through your gradient descent}

\item{step.size}{The size of the step to take during gradient descent}
}
\value{
A unscaled weight matrix [max.iterations x p]
}
\description{
LMSquareLossIterations
}
\examples{
x.unsc.mat <- data[,-1]
y.vec <-data[,1]
max.iterations=10
step.size=0.35
W.train <- LMLogisticLossIterations(X.train, Y.train, max.iterations, step.size)
cbind(1, X.mat) \%*\% W.mat
}
