% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_Models.r
\name{LMSquareLossL2}
\alias{LMSquareLossL2}
\title{L2 Square loss}
\usage{
LMSquareLossL2(X.scaled.mat, y.vec, penalty, opt.thresh,
  initial.weight.vec, step.size)
}
\arguments{
\item{X.scaled.mat}{numeric scaled input feature matrix [n x p]}

\item{y.vec}{numeric input label vector [n]}

\item{penalty}{numeric value for the penalty on weights}

\item{opt.thresh}{numeric value compare for convergence}

\item{initial.weight.vec}{numeric 0 vector [p]}

\item{step.size}{numeric learning rate}
}
\value{
numeric scaled optimal weight vector for the given penalty [p]
}
\description{
This function attempts to minimize the Loss function [f(x) - y]^2. Where f(x) is the resulting prediction values of x and y are truth values.
}
