% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_Models.r
\name{LMSquareLossL2CV}
\alias{LMSquareLossL2CV}
\title{L2 Square Loss CV}
\usage{
LMSquareLossL2CV(X.mat, y.vec, fold.vec = NULL, penalty.vec)
}
\arguments{
\item{X.mat}{numeric input feature matrix [n x p]}

\item{y.vec}{numeric input label vector [n]}

\item{fold.vec}{numeric fold ids for each observation in X.mat [n]}

\item{penalty.vec}{numeric vector of decending penalties for weights [m]}
}
\value{
list containing the mean validation loss, mean training loss vector, pentalty vector, optimal penalty, optimal weight vector using optimal penalty, predict function
}
\description{
Cross validation using L2 Square Loss regularization
}
\examples{
data(ozone, package="ElemStatLearn")
X.mat <- ozone[, 2:ncol(ozone)]
y.vec <- ozone[, "ozone"]
fold.vec <- sample(rep(1:4, l=nrow(X.mat)))
penalty.vec <- seq(10, 1, by=-0.5)
LMSquareLossL2CV(X.mat, y.vec, fold.vec, penalty.vec)
}
