% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeRegression.R
\docType{data}
\name{RidgeRegression}
\alias{RidgeRegression}
\title{RidgeRegression}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RidgeRegression
}
\description{
RidgeRegression
}
\examples{

library(linearmodels)

x <- matrix(runif(100), ncol = 2)
y <- matrix(2*x[,1] + 3*x[, 2] + rnorm(50), ncol = 1)

ridge_regression <- RidgeRegression$new(lambda = 0.1)
ridge_regression$fit(x, y)
ridge_regression$predict(x)
}
\keyword{datasets}
