% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarlyStopping.R
\name{LMLogisticLossEarlyStoppingCV}
\alias{LMLogisticLossEarlyStoppingCV}
\title{LMLogisticLossEarlyStoppingCV function}
\usage{
LMLogisticLossEarlyStoppingCV(X.mat, y.vec, fold.vec, max.iterations)
}
\arguments{
\item{X.mat}{numeric train featrue matrix [n x p]}

\item{y.vec}{numeric train label vector [n]}

\item{fold.vec}{numeric fold id vector [n]}

\item{max.iterations}{scalar integer greater than 1, max number of iterations}
}
\value{
mean.validation.loss

mean.train.loss.vec

selected.steps

weight.vec  the weight vector found by using gradient descent with selected.steps on the whole training data set.

predict(testX.mat) a function that takes a test features matrix and returns a vector of predictions
}
\description{
R function that
}
\examples{
data(zip.train, package="ElemStatLearn")
i01 <- which(zip.train[,1] \%in\% c(0,1))
train.i <- i01[1:5]
test.i <- i01[6]
x <- zip.train[train.i, -1]
y <- zip.train[train.i, 1]
testx <- zip.train[test.i, -1]
LMLogisticLossIterations(x ,y, testx , as.integer(3))
zip.train[test.i, 1]

}
