% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarlyStopping.R
\name{LMSquareLossIterations}
\alias{LMSquareLossIterations}
\title{LMSquareLossIterations function}
\usage{
LMSquareLossIterations(X.mat, y.vec, max.iterations, step.size)
}
\arguments{
\item{X.mat}{numeric train featrue matrix [n x p]}

\item{y.vec}{numeric train label vector [n]}

\item{step.size}{scalar integer, step size}

\item{max.iteratons}{scalar integer greater than 1, max number of iterations}
}
\value{
matrix of weight vectors
}
\description{
R function that wraps the C++ code.
}
\examples{
data(prostate, package="ElemStatLearn")
x <- prostate[1, ]
y <- prostate[, 'lpsa']
LMSquareLossIterations(x ,y, as.integer(12), as.integer(0.5))

}
